requirements.txt
# Core Requirements
# Python 3.7+

# LLM API Clients (choose one or both)
openai>=1.0.0          # For OpenAI GPT models
anthropic>=0.18.0      # For Anthropic Claude models

# Speech Recognition (Optional)
speechrecognition>=3.10.0
pyaudio>=0.2.11        # Note: May require system audio libraries

# Text-to-Speech (Optional)
pyttsx3>=2.90          # Cross-platform TTS

# Web Server
flask>=3.0.0           # Web framework for hosting
flask-cors>=4.0.0      # CORS support for API

# Environment Variables
python-dotenv>=1.0.0   # Load environment variables from .env file

# Note: For PyAudio on Windows, you may need to install it separately:
# pip install pipwin
# pipwin install pyaudio

# No additional dependencies required - uses standard library for:
# - CSV handling
# - JSON processing
# - File I/O
# - Date/time operations

